First real commit.
This is the first commit of actual code. The csv2markdown.pl file is the perl
script that does the conversion. The example.csv file is, as the name implies,
an example CSV for testing.

The csv2markdown.pl file isn't working yet. It appears I don't understand
multidimensional arrays in Perl. Probably switch to hashes because it looks like
that's pretty much how they're done anyway, based on a cursory glance at google.