HASHIFIED, NOT THERE YET
The outer array is now a hash. References are being passed and used now. I'm
getting closer to working, but there's still issues. I found one of the big ones
which was that double quote characters were included in an array to track where
the quotes occur, then checked in an if conditional with a simple "eq" check.

The sizes aren't right yet, though. The last column is 1 when it should be
the longest. It may also be extraneous.

More to come.