ALMOST THERE!
I've got most of the table showing up, now. What isn't happening is I'm not
seeing the last column of lines that just end in a comma, but the ones that end
in double quotes are working just fine. I'll have to figure out what's up with
that. Beyond that, there's a whole lot of debugging messages to clean up.