SUPER CLOSE NOW. ALL COLUMNS PRINT. QUOTE LINES HAVE EXTRA!
It's so close I can taste it! I got all the columns printing. For some reason,
though, despite my best efforts, the quote lines have an extra column as
evidenced by the extra pipe character printed at the end. Also, I have a debug
statement that saves the detected size of the lineRef array passed to the
printLine() function and it keeps showing the number as 5, when it should be 4.
So, the problems are a little deeper than stripping extra commas. That, or my
effort to do that failed. No time to debug now. Gotta just do a quick check-in.