GOT IT. IT WAS THE STUPID NEWLINE.
It turns out that the lesson is to always use chomp() when you pull in lines
with the standard angle-bracket operator. The split function will normally
ignore end-of-line delimiter. Thing is, it's expecting a null character to end
the line, not a newline. So, I had the newline present because the lack of
chomp(), and on top of that, my conditional to check for an extra element after
the final delimiter was looking for a null string, not one that contained a
newline. So, I got that mysterious extra element. But I finally got it licked!

So, once I got that, there was a remaining issue with the final part of the
header breaking line not printing. It turned out that was caused by a bad
conditional in my outer for loop. It was checking for less than the size of an
array minus one, when it should have been checking for less than the size of the
array, period. Beyond that, I decided I wanted the padding to lead all the
lines, as well as end them.

Anyway, CSV files now convert to MarkDown tables!